% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornado.R
\name{tornado}
\alias{tornado}
\alias{tornado_sample}
\alias{tornado_plot}
\title{Tornado sampling and plotting function}
\usage{
tornado_sample(
  n,
  fun,
  distributions,
  ...,
  baseline = NULL,
  output_name = ".result"
)

tornado_plot(x, ..., type = c("jitter", "maxmin"), nbreaks = 6, xlab = "value")
}
\arguments{
\item{n}{Number of samples.}

\item{fun}{A vectorised function.}

\item{distributions}{Distributions created with the
\href{https://alexpghayes.github.io/distributions3/}{distributions3} package that
correspond to the named parameters of \code{fun}.}

\item{...}{Not currently used.}

\item{baseline}{NULL or a scalar numeric value. If not NULL then this value
is used as the comparitor when producing the latter tornado plots.}

\item{output_name}{The name to use of the result column in the output
tibbles. This must not correspond to any of the parameter names of \code{fun}.}

\item{x}{Output from \code{tornado_sample()}.}

\item{type}{Character string. Either 'jitter' or 'maxmin'. If 'jitter' then
we simply plot the sample results against the corresponding distribution that
varies. Otherwise, we plot the corresponding maximum and minimum values as a
split bar chart.}

\item{nbreaks}{Passed to \code{scales::breaks_extended()}.}

\item{xlab}{The label for the x-axis.}
}
\value{
For \code{tornado_sample()} a tibble with class \strong{tornado_samples} and an
additional attribute that gives the baseline value calculated when setting
each argument to it's distributions mean value (or an optional,
user-supplied, value). For \code{tornado_plot()}, a \code{ggplot2} object
}
\description{
A small wrapper around lapply that loops over a functions parameters,
sampling one whilst fixing the others. \code{tornado_sample()} just wraps a
simple univariate sensitivity analysis, ensuring the output is formatted in a
usable way for the namesake plot function.
}
\examples{

# User inputs ----------------------------------------------------------

# Function with desired arguments
# NOTE: MUST be vectorised by the user and return a scalar value
fun <- function(cost_per_day, days, discount) cost_per_day * days * discount

# A distribution (from distributions3) for each function argument
distributions <- list(
    cost_per_day = distributions3::Gamma(shape = 9, rate = 0.5),
    days         = distributions3::Gamma(shape = 5, rate = 1),
    discount     = distributions3::Beta(alpha = 2, beta = 40)
)

# Number of samples to take for each argument
samples <- 1000

# Package functionality ------------------------------------------------

# generate the samples for each parameter
dat <- tornado_sample(samples, fun, distributions)

# jitter plot
tornado_plot(dat)

# plot of the extremes (akin to the more traditional plots)
tornado_plot(dat, type = "maxmin")

}
